extract add command
